requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
discord-components==1.0.1
discord.py==1.7.3
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
giphy-client==1.0.0
html5lib==1.1
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
keyboard==0.13.5
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.0.2
prompt-toolkit==3.0.18
psutil==5.8.0
pycodestyle==2.7.0
pyenchant==3.2.1
pygame==2.0.1
Pygments==2.9.0
pynput==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3